# documentazione/swagger-openapi.yaml
openapi: 3.0.1
info:
  title: OSP API
  description: Gestionale Occupazione Suolo Pubblico Bari
  version: '1.0'
servers:
  - url: http://localhost:8080/api/osp
    description: Generated server url
paths:
...
  /template-management/template:
    get:
      tags:
        - template-controller
      operationId: getTemplate
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateSimplifiedDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
    put:
      tags:
        - template-controller
      operationId: updateTemplate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateDTO'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
    post:
      tags:
        - template-controller
      operationId: insertTemplate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TemplateDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /richiesta-integrazione-management/richiesta-rettifica-date:
    put:
      tags:
        - richiesta-integrazione-controller
      operationId: richiestaRettificaDate
      parameters:
        - name: idPratica
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: idUtente
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: notaAlCittadino
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RichiestaIntegrazioneDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
  /richiesta-parere-management/richiesta-parere/verifica-formale:
    post:
      tags:
        - richiesta-parere-controller
      operationId: insertRichiestaParereFromVerificaFormale
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaParereDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RichiestaParereDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /richiesta-parere-management/richiesta-parere/ulteriori-pareri:
    post:
      tags:
        - richiesta-parere-controller
      operationId: insertUlterioriRichiestePareri
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UlterioriPareriInsertRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RichiestaParereDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /richiesta-integrazione-management/richiesta-integrazione:
    post:
      tags:
        - richiesta-integrazione-controller
      operationId: insertRichiestaIntegrazione
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaIntegrazioneDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RichiestaIntegrazioneDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
  /template-management/template/{id}:
    get:
      tags:
        - template-controller
      operationId: getTemplate_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TemplateDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /template-management/template/tipo:
    get:
      tags:
        - template-controller
      operationId: getTemplateByTipo
      parameters:
        - name: idTipoTemplate
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TemplateDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /template-management/template/elaborato:
    get:
      tags:
        - template-controller
      operationId: getTemplateElaborato
      parameters:
        - name: idPratica
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: idTipoTemplate
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: notaParere
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TemplateDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /stats/caricaDashboardKibana:
    get:
      tags:
        - statistiche-controller
      operationId: getTemplate_2
      parameters:
        - name: municipio
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DashboardDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
  /stats/caricaDashboardKibana/{municipio}:
    get:
      tags:
        - statistiche-controller
      operationId: getTemplate_3
      parameters:
        - name: municipio
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DashboardDTO'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: object
      security:
        - jwt: []
...
components:
  schemas:
...
    TemplateDTO:
      required:
        - fileTemplate
        - mimeType
        - nomeFile
      type: object
      properties:
        id:
          type: integer
          format: int32
        tipoTemplate:
          $ref: '#/components/schemas/TypologicalDTO'
        nomeFile:
          maxLength: 255
          minLength: 0
          type: string
        mimeType:
          maxLength: 255
          minLength: 0
          type: string
        dataInserimento:
          type: string
          format: date-time
        dataModifica:
          type: string
          format: date-time
        utenteModifica:
          $ref: '#/components/schemas/UtenteDTO'
        fileTemplate:
          type: string
    TypologicalDTO:
      required:
        - descrizione
        - descrizioneEstesa
        - id
      type: object
      properties:
        id:
          type: integer
          format: int32
        descrizione:
          type: string
        descrizioneEstesa:
          type: string
    UtenteDTO:
      required:
        - codiceFiscale
        - dateCreated
        - email
        - idGruppo
        - username
      type: object
      properties:
        id:
          type: integer
          format: int64
        idGruppo:
          type: integer
          format: int32
        idsMunicipi:
          uniqueItems: true
          type: array
          items:
            type: integer
            format: int32
        username:
          maxLength: 30
          minLength: 3
          type: string
        nome:
          maxLength: 50
          minLength: 0
          type: string
        cognome:
          maxLength: 50
          minLength: 0
          type: string
        sesso:
          type: string
          enum:
            - M
            - F
        dataDiNascita:
          type: string
          format: date
        luogoDiNascita:
          maxLength: 50
          minLength: 0
          type: string
        provinciaDiNascita:
          maxLength: 50
          minLength: 0
          type: string
        codiceFiscale:
          maxLength: 16
          minLength: 11
          type: string
        ragioneSociale:
          maxLength: 50
          minLength: 0
          type: string
        email:
          maxLength: 50
          minLength: 0
          type: string
        numTel:
          maxLength: 15
          minLength: 0
          type: string
        enabled:
          type: boolean
        dateCreated:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
        flagEliminato:
          type: boolean
        dataEliminazione:
          type: string
          format: date-time
        uoId:
          maxLength: 255
          minLength: 0
          type: string
        indirizzo:
          maxLength: 255
          minLength: 0
          type: string
    IntegrazioneDTO:
      required:
        - idRichiestaIntegrazione
        - idUtenteIntegrazione
      type: object
      properties:
        id:
          type: integer
          format: int64
        idRichiestaIntegrazione:
          type: integer
          format: int64
        idUtenteIntegrazione:
          type: integer
          format: int64
        motivoIntegrazione:
          maxLength: 512
          minLength: 0
          type: string
        codiceProtocollo:
          type: string
        numeroProtocollo:
          type: string
        anno:
          type: string
        dataProtocollo:
          type: string
          format: date-time
        dataInserimento:
          type: string
          format: date-time
        dataInizioOccupazione:
          type: string
          format: date
        oraInizioOccupazione:
          $ref: '#/components/schemas/LocalTime'
        dataScadenzaOccupazione:
          type: string
          format: date
        oraScadenzaOccupazione:
          $ref: '#/components/schemas/LocalTime'
        nomeCittadinoEgov:
          type: string
        cognomeCittadinoEgov:
          type: string
        cfCittadinoEgov:
          type: string
        statoPratica:
          $ref: '#/components/schemas/TypologicalDTO'
        originEgov:
          type: boolean
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32
    RichiestaIntegrazioneDTO:
      required:
        - idPratica
        - idUtenteRichiedente
      type: object
      properties:
        id:
          type: integer
          format: int64
        idPratica:
          type: integer
          format: int64
        idUtenteRichiedente:
          type: integer
          format: int64
        idStatoPratica:
          type: integer
          format: int32
        codiceProtocollo:
          maxLength: 100
          minLength: 0
          type: string
        dataProtocollo:
          type: string
          format: date-time
        tipoRichiesta:
          type: string
          enum:
            - PROCEDURA_DINIEGO
            - RETTIFICA_DATE
        motivoRichiesta:
          maxLength: 512
          minLength: 0
          type: string
        dataInserimento:
          type: string
          format: date-time
        dataScadenza:
          type: string
          format: date-time
        integrazione:
          $ref: '#/components/schemas/IntegrazioneDTO'
        flagEsenzioneMarcaDaBollo:
          type: string
        idAllegato:
          type: integer
          format: int64
...
    ParereDTO:
      required:
        - idRichiestaParere
        - idUtenteParere
      type: object
      properties:
        id:
          type: integer
          format: int64
        idRichiestaParere:
          type: integer
          format: int64
        idUtenteParere:
          type: integer
          format: int64
        nota:
          maxLength: 255
          minLength: 0
          type: string
        esito:
          type: boolean
        codiceProtocollo:
          type: string
        dataProtocollo:
          type: string
          format: date-time
        dataInserimento:
          type: string
          format: date-time
        flagCompetenza:
          type: boolean
...
    RichiestaParereDTO:
      required:
        - idPratica
        - idUtenteRichiedente
      type: object
      properties:
        id:
          type: integer
          format: int64
        idPratica:
          type: integer
          format: int64
        idUtenteRichiedente:
          type: integer
          format: int64
        idStatoPratica:
          type: integer
          format: int32
        idGruppoDestinatarioParere:
          type: integer
          format: int32
        codiceProtocollo:
          type: string
        dataProtocollo:
          type: string
          format: date-time
        notaRichiestaParere:
          maxLength: 512
          minLength: 0
          type: string
        dataInserimento:
          type: string
          format: date-time
        flagInseritaRisposta:
          type: boolean
        parere:
          $ref: '#/components/schemas/ParereDTO'
        riabilitaEsenzioneMarcaDaBollo:
          type: boolean
...
    UlterioriPareriInsertRequest:
      required:
        - idGruppiDestinatariPareri
        - idPratica
      type: object
      properties:
        idPratica:
          type: integer
          format: int64
        idGruppiDestinatariPareri:
          type: array
          items:
            type: integer
            format: int32
...
    TemplateSimplifiedDTO:
      required:
        - mimeType
        - nomeFile
      type: object
      properties:
        id:
          type: integer
          format: int32
        tipoTemplate:
          $ref: '#/components/schemas/TypologicalDTO'
        nomeFile:
          maxLength: 255
          minLength: 0
          type: string
        mimeType:
          maxLength: 255
          minLength: 0
          type: string
        dataInserimento:
          type: string
          format: date-time
        dataModifica:
          type: string
          format: date-time
        utenteModifica:
          $ref: '#/components/schemas/UtenteDTO'
    DashboardDTO:
      type: object
      properties:
        dashboard:
          type: string
...
  securitySchemes:
    jwt:
      type: http
      in: header
      scheme: Bearer